Access array data correctly
Multidimensional arrays in C are a lot harder than in Java. However I
seem to have solved it by learning about how C stores the
multidimensional array in memory (just as a long sequence) and then
doing pointer arithmics to get the correct value. So now the arrays
should work.