Put more logic into sendBuffer
If sendbuffer is allowed to be a bit more fleshed out it is much easier
to keep away from global variables. So I'm gonna go ahead and do that.
Let the sendBuffer function hold onto state for the buffer etc. Might
make it easier to expand to NKRO in the future if I find 6KRO to be
limiting. Doubt that though.