More debugging
Now we can properly read if we have a changed state and what keys are
pressed. Trying to figure out why the layer does not get updated
correctly.

And we are not sending any keys. It is most likely the same issue so
solve one and apply the solution to the other.