Correctrly read state
The order of the row pins was wrong. This is why the layers were off.
Working on building the buffer now. It seems we are not building it for
some reason. We can see that we should be sending metas, but not any
other keys. Need to figure out what's going on here.