Add fallthrough
If you press a key that is on a higher layer, but that position does not
have a value, fall through to the layer beneth it and get that key.

Example: Your press Raise. This takes you to layer 2. You press a key
that is not bound to anything. We go down to layer 1 and see if we have
anything bound there. If we don't find anything we go to layer 0 and
check. If we still can't find anything, nothing happens.

This makes it possible to press F3+B without having to specify B on the
same layer as F3.